Move auto-registration to explicit install packages, register plugin interfaces
