+++ container/crio/install/install.go
@@ -12,22 +12,18 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package containerd
+// The install package registers crio.NewPlugin() as the "crio" container provider when imported
+package install
 
 import (
 	"github.com/google/cadvisor/container"
-	"github.com/google/cadvisor/fs"
-	info "github.com/google/cadvisor/info/v1"
-	"github.com/google/cadvisor/watcher"
+	"github.com/google/cadvisor/container/crio"
 	"k8s.io/klog"
 )
 
 func init() {
-	err := container.RegisterPlugin("containerd", func(factory info.MachineInfoFactory, fsInfo fs.FsInfo, includedMetrics container.MetricSet) (watcher.ContainerWatcher, error) {
-		err := Register(factory, fsInfo, includedMetrics)
-		return nil, err
-	})
+	err := container.RegisterPlugin("crio", crio.NewPlugin())
 	if err != nil {
-		klog.Fatalf("Failed to register containerd plugin: %v", err)
+		klog.Fatalf("Failed to register crio plugin: %v", err)
 	}
 }
